Added masonry-grid and art-card components
